% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TravelTimeReliability.R
\name{TravelTimeReliability}
\alias{TravelTimeReliability}
\title{Calculate travel time reliability measures
\code{TravelTimeReliability} calculates freeway and arterial travel time index
by congestion level}
\usage{
TravelTimeReliability(L)
}
\arguments{
\item{L}{A list containing data defined by the module specification.}
}
\value{
A list containing data produced by the function consistent with the
module specifications.
}
\description{
This function uses the average free flow speed (congestion level none) and the
average speeds during the remaining four congestion levels (mod, hvy, sev, ext)
and calculates travel time index as a ratio of the congested speed and free flow
speed
}
