% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TravelTimeReliability.R
\docType{data}
\name{TravelTimeReliabilitySpecifications}
\alias{TravelTimeReliabilitySpecifications}
\title{Specifications list for TravelTimeReliability module}
\format{
A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}
}
\source{
TravelTimeReliability.R script.
}
\usage{
TravelTimeReliabilitySpecifications
}
\description{
A list containing specifications for the TravelTimeReliability module.
}
\keyword{datasets}
